+++ src/events/DefineEntryTypesEvent.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace mmikkel\childme\events;
+
+use yii\base\Event;
+
+class DefineEntryTypesEvent extends Event
+{
+
+    /**
+     * @var string
+     */
+    public $section;
+
+    /**
+     * @var array 
+     */
+    public $entryTypes = [];
+
+}